+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 * add a panel to select extra icons to activate.
 
 ### 0.8 (WIP)
-* support `.mailmap`, `version` (with `.md`, `.adoc`, `.txt` or no extension) files.
 * reworked `*.iml`, `NOTICE` files.
+* support `.mailmap`, `*.kdbx` (KeePass), `version` (with `.md`, `.adoc`, `.txt` or no extension) files.
+* support `*.pem`, `*.crt`, `*.ca-bundle`, `*.cer`, `*.p7b`, `*.p7s`, `*.pfx` certificate files.
 
 ### 0.7 (2018/09/06)
 * support `vagrantfile`, `*.iml`, `.gitattributes`, `.gitignore`, `.editorconfig`, `berksfile`, `berksfile.lock`, `dockerfile`, `docker-compose.yml` files.